switch weight type to float + regex the crap out of it

+++ db/migrate/20140405193714_add_user_attributes.rb
@@ -1,16 +1,16 @@
 class AddUserAttributes < ActiveRecord::Migration
   def change
     add_column :users, :date_of_birth, :string
-    add_column :users, :weight, :integer
-    add_column :users, :ideal_weight, :integer
+    add_column :users, :weight, :float
+    add_column :users, :ideal_weight, :float
     add_column :users, :do_sport, :boolean
     add_column :users, :would_do_sport, :boolean
   end
 
   def self.up
     add_column :users, :date_of_birth, :string
-    add_column :users, :weight, :integer
-    add_column :users, :ideal_weight, :integer
+    add_column :users, :weight, :float
+    add_column :users, :ideal_weight, :float
     add_column :users, :do_sport, :boolean
     add_column :users, :would_do_sport, :boolean
   end
